Small patches here and there.
I tacked on a couple of things, like decrementing the potion count when
you dring a potion. Also, I changed the constructors to use the setter
methods; I think Bennett considers that to be the better style. This
also gave a reason to have the setter in Weapon available. Let me know
what you think.